# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BEYOND-MOBILE-BACKEND",
        "title": "BEYOND-MOBILE-BACKEND",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/internal/verihubs/verify-ktp": {
            "post": {
                "description": "Endpoint for verify user KTP for KYC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KYC"
                ],
                "summary": "KYC Verify KTP User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "KTP Verify Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.KycRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RegisterEmailAlreadyUsed"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/v1/internal/verihubs/verify-passport": {
            "post": {
                "description": "Endpoint for verify user Passport for KYC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KYC"
                ],
                "summary": "KYC Verify Passport User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "KTP Verify Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.KycRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RegisterEmailAlreadyUsed"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/kyc/register-kyc-ktp": {
            "post": {
                "description": "Add new user KTP for KYC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KYC"
                ],
                "summary": "KYC Save KTP User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "KTP Save Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.KTPrequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RegisterEmailAlreadyUsed"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/kyc/register-kyc-passport": {
            "post": {
                "description": "Add new user Passport for KYC",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KYC"
                ],
                "summary": "KYC Save Passport User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Passport Save Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.PassportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RegisterEmailAlreadyUsed"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend-mobile-api_model_dto_request.DetailCurrentAddress": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "full_address": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.KTPrequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "current_address": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.DetailCurrentAddress"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "martial_status": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "occupation": {
                    "type": "string"
                },
                "place_of_birth": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "backend-mobile-api_model_dto_request.KycRequest": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.PassportRequest": {
            "type": "object",
            "properties": {
                "current_address": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.DetailCurrentAddress"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "date_of_expired": {
                    "type": "string"
                },
                "date_of_issue": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "passport_no": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "place_of_birth": {
                    "type": "string"
                },
                "place_of_issue": {
                    "type": "string"
                },
                "resi_number": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.BasicSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "string",
                    "example": "00"
                }
            }
        },
        "backend-mobile-api_model_dto_swagger.CommonError": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "time out"
                },
                "message": {
                    "type": "string",
                    "example": "server is busy"
                },
                "status_code": {
                    "type": "string",
                    "example": "119"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.RegisterEmailAlreadyUsed": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "deference confirmed pin"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request payload"
                },
                "status_code": {
                    "type": "string",
                    "example": "110"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.UserNotFound": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "user not found"
                },
                "status_code": {
                    "type": "string",
                    "example": "117"
                }
            }
        },
...
    }
}